+++ changelog.md
@@ -3,3 +3,4 @@
 - Added Pest Highlighter + Tracers
 - Detect Glacite Corpses through walls
 - Make Livid Solver better
+- ghost blocc